+++ app/_layout.tsx
@@ -40,6 +40,7 @@ const RootLayout = () => {
       <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
       <Stack.Screen name="(auth)" options={{ headerShown: false }} />
       <Stack.Screen name="index" options={{ headerShown: false }} />
+      <Stack.Screen name="achiverment/index" options={{ headerShown: false }} />
     </Stack>
   );
 };
